Restructure code into package
