+++ airquality/enums.py
